# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Testing Shop API",
    "version": "1.0.0",
    "description": "Demo e-shop backend with fault injection for API/UI/Unit testing scenarios."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://{apiHost}",
      "description": "Deployment host",
      "variables": {
        "apiHost": {
          "default": "example.azurecontainerapps.io"
        }
      }
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "CartSessionHeader": {
        "in": "header",
        "name": "X-Cart-Session",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Guest cart session key used for cart and checkout operations."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "Money": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number"
          },
          "currencyCode": {
            "type": "string",
            "example": "CZK"
          }
        },
        "required": [
          "amount",
          "currencyCode"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "inStock": {
            "type": "integer"
          },
          "active": {
            "type": "boolean"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "inStock",
          "active",
          "price"
        ]
      },
      "StorefrontCatalogResponse": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "priceBounds": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number"
              },
              "max": {
                "type": "number"
              },
              "currencyCode": {
                "type": "string"
              }
            },
            "required": [
              "min",
              "max",
              "currencyCode"
            ]
          }
        },
        "required": [
          "products",
          "priceBounds"
        ]
      },
...
    }
  },
...
  "paths": {
...
    "/admin/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List all products for admin editing.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Product list for admin.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create product as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "inStock": {
                    "type": "integer"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "price": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "number"
                      },
                      "currencyCode": {
                        "type": "string",
                        "example": "CZK"
                      }
                    },
                    "required": [
                      "amount",
                      "currencyCode"
                    ]
                  }
                },
                "required": [
                  "name",
                  "description",
                  "inStock",
                  "active",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product by id as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "inStock": {
                    "type": "integer"
                  },
                  "active": {
                    "type": "boolean"
                  },
                  "price": {
                    "type": "object",
                    "properties": {
                      "amount": {
                        "type": "number"
                      },
                      "currencyCode": {
                        "type": "string",
                        "example": "CZK"
                      }
                    },
                    "required": [
                      "amount",
                      "currencyCode"
                    ]
                  }
                },
                "required": [
                  "name",
                  "description",
                  "inStock",
                  "active",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product by id as admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted."
          },
          "400": {
            "description": "Invalid id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/checkout/bank-transfer": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Finalize order via bank transfer flow and send confirmation email.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartSessionHeader"
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "customerFirstName": {
                    "type": "string"
                  },
                  "customerLastName": {
                    "type": "string"
                  },
                  "customerPhone": {
                    "type": "string"
                  },
                  "addressLine1": {
                    "type": "string"
                  },
                  "addressLine2": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "postalCode": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  }
                },
                "required": [
                  "customerEmail",
                  "customerFirstName",
                  "customerLastName",
                  "customerPhone"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created with bank transfer details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Cart or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/checkout/gateway/init": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Initialize payment gateway checkout and create/reuse pending order.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CartSessionHeader"
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "customerFirstName": {
                    "type": "string"
                  },
                  "customerLastName": {
                    "type": "string"
                  },
                  "customerPhone": {
                    "type": "string"
                  }
                },
                "required": [
                  "customerEmail",
                  "customerFirstName",
                  "customerLastName",
                  "customerPhone"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pending gateway order initialized.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/checkout/gateway/{orderId}/mock-pay": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Simulate gateway capture for a pending order.",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mock gateway result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get storefront product catalog with filtering/sorting/localization.",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search term for name/description."
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "en",
                "cs"
              ]
            },
            "description": "Storefront language."
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "name-asc",
                "name-desc",
                "price-asc",
                "price-desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "priceMin",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "priceMax",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Storefront catalog response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorefrontCatalogResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}